Jurizdictii (ACL) - adica permisiuni pe date, facute programatic cu if!

+++ spring-one/src/main/java/victor/training/spring/web/controller/TrainingController.java
@@ -6,8 +6,13 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import victor.training.spring.web.controller.dto.TrainingDto;
+import victor.training.spring.web.domain.Training;
+import victor.training.spring.web.domain.User;
+import victor.training.spring.web.repo.TrainingRepo;
+import victor.training.spring.web.repo.UserRepo;
 import victor.training.spring.web.service.TrainingService;
 
 @RestController
@@ -45,8 +50,26 @@ public class TrainingController {
    @DeleteMapping("{id}")
    @PreAuthorize("hasRole('ADMIN')") // anotation-based authorization.
    public void deleteTrainingById(@PathVariable Long id) {
+      Training training = trainingRepo.findById(id).orElseThrow(() -> new RuntimeException("Draga hackere, n-ai nimerit id-ul. Mai baga o fisa!?"));
+
+      Long teacherId = training.getTeacher().getId();
+      String currentUsername = SecurityContextHolder.getContext().getAuthentication().getName();
+
+      User user = userRepo.findByUsername(currentUsername);
+
+      if (!user.getManagedTeacherIds().contains(teacherId)) {
+         throw new IllegalArgumentException("N-ai voie!");
+      }
+
       trainingService.deleteById(id);
    }
+
+   @Autowired
+   private TrainingRepo trainingRepo;
+   @Autowired
+   private UserRepo userRepo;
+
+
 
    @PostMapping
    public void createTraining(@RequestBody TrainingDto dto) throws ParseException {
